Organize unit/integration/e2e tests
