+++ jest-all.json
@@ -2,7 +2,7 @@
   "moduleFileExtensions": ["js", "json", "ts"],
   "rootDir": ".",
   "testEnvironment": "node",
-  "testRegex": ".spec-int.ts$",
+  "testRegex": ".spec-?(int|e2e)?.ts$",
   "transform": {
     "^.+\\.(t|j)s$": "ts-jest"
   }
