+++ jest-e2e.json
@@ -2,7 +2,7 @@
   "moduleFileExtensions": ["js", "json", "ts"],
   "rootDir": ".",
   "testEnvironment": "node",
-  "testRegex": ".e2e-spec.ts$",
+  "testRegex": ".spec-e2e.ts$",
   "transform": {
     "^.+\\.(t|j)s$": "ts-jest"
   }
