+++ package.json
@@ -7,21 +7,20 @@
   "license": "UNLICENSED",
   "scripts": {
     "build": "nest build",
-    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "nest start",
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
+    "format": "prettier --write \"src/**/*.ts\"",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-    "test": "jest",
-    "test:cov": "jest --coverage",
-    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:int": "npm run db:restart:test && dotenv -e .env.test -- jest --config ./jest-int.json --runInBand  --no-cache",
-    "test:e2e": "npm run db:restart:test && dotenv -e .env.test -- jest --config ./test/jest-e2e.json",
-    "test:all": "npm run test && npm run test:int && npm run test:e2e",
-    "test:watch": "jest --watch",
+    "test:unit": "jest",
+    "test:int": "npm run db:restart:test && dotenv -e .env.test -- jest --config ./jest-int.json --runInBand --no-cache",
+    "test:e2e": "npm run db:restart:test && dotenv -e .env.test -- jest --config ./jest-e2e.json --runInBand --no-cache",
+    "test:all": "npm run db:restart:test && dotenv -e .env.test -- jest --config ./jest-all.json --runInBand --no-cache",
+    "test:unit:watch": "jest --watch",
     "test:int:watch": "npm run test:int -- --watch",
     "test:e2e:watch": "npm run test:e2e -- --watch",
+    "test:all:watch": "npm run test:all -- --watch",
     "db:restart:dev": "npm run docker:up:dev && npm run prisma:migrate:dev",
     "db:restart:test": "npm run docker:up:test && npm run prisma:migrate:test",
     "prisma:migrate:dev": "dotenv -e .env.dev -- prisma migrate deploy",
