+++ src/app/app.controller.spec-e2e.ts
@@ -1,7 +1,7 @@
-import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
+import { Test, TestingModule } from '@nestjs/testing';
 import * as request from 'supertest';
-import { AppModule } from './../src/app.module';
+import { AppModule } from './app.module';
 
 describe('AppController (e2e)', () => {
   let app: INestApplication;
