+++ src/tax-report/tax-report.controller.spec-int.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { Prisma } from '@prisma/client';
-import { AppModule } from '../app.module';
+import { AppModule } from '../app/app.module';
 import { PrismaService } from '../prisma/prisma.service';
 import { TaxReportController } from './tax-report.controller';
